Fix enums
