+++ CMakeLists.txt
@@ -11,7 +11,8 @@ set(CMAKE_CXX_FLAGS_DEBUG_INIT "-g -fsanitize=address -fsanitize=undefined")
 set(CMAKE_CXX_FLAGS_RELEASE_INIT "-O2")
 
 file(GLOB_RECURSE HEADERS "include/*.h" "third_party/*.hpp")
-file(GLOB_RECURSE SOURCES "src/*.cc" )
+file(GLOB_RECURSE SOURCES "src/*.cc"  "src/autocomplete/*.cc")
+
 
 find_package(Qt5Core CONFIG REQUIRED)
 find_package(Qt5Widgets CONFIG REQUIRED)
